split music.js in smaller files
